fix(Pointer): update pointer position in fixed update
The Straight Pointer Renderer position was being updated using the
Transform Follow script and this was defaulting to using the
OnPreRender moment to update position. This had a knock on effect
that the pointer Object Interactor position was being updated in a
Fixed Update but this meant that the Object Interactor position was
not correctly updated to the correct pointer position as the pointer
position was always slightly slower due to being on the OnPreRender
setting.

The solution is to ensure the Pointer position is updated via
FixedUpdate as well to ensure the position of the pointer is correct
for when the Object Interactor is updated in the Fixed Update routine.